Added HomeScreenUtils.kt - extracted composable funs into there , Refactored HomeScreen.kt
